Get the left side of the flavors dialog working, pixel-perfect fractional scale

+++ src/puppet/java/com/unascribed/sup/opengl/util/GL.java
@@ -81,7 +81,7 @@ public class GL extends GL12 {
 	}
 	
 
-	public static void drawCircle(float x, float y, int dia) {
+	public static void drawCircle(float x, float y, float dia) {
 		State s = state.get();
 		int list = s.circleList;
 		if (list == 0) {
